Avoid DOM updates in constructors with this.attach
The new attach method is intended for code that makes changes to the
DOM but should only be run once. Moving code to this method in each
class helps to simplify the constructor, which should be focused on
internal work to initialize the object and not making external changes
to the DOM. It also helps to support future behaviour where attachment
to a DOM element may not be possible at the time of object construction.