Explictly select initial options for the app
While the initial options for the layout and schema are highlighted in
the radio taxele, we should also make sure the associated handlers are
invoked with these initial settings to ensure they are setup similarly.

This highlights an ordering constraint where layout changes should
happen after the initial data received trigger. Rather than add more
logic to the layout handler to cope with this scenario, it makes more
sense to extract the call to the application level so that the
constraint is clearly visible at this point in the code.