Simplify taxplay into a single update stage
The taxplay enter and exit stages aren't necessary as only one item of
data is being rendered, so there's no associated data binding to
process. Rather than calling both the attach and the update taxele
methods in the same handler, it makes more sense to combine them as they
are never used in isolation - collapsing the stages made this clear.