Replace taxplay form with explicit set/commit
Previously the taxplay would use normal form features to achieve its
save/reset behaviour, but the taxeles used in the taxplay are also used
in the source/layout selectors, where a form context is not appropriate.

Given that each editable taxele is constructed with a data node, it
makes sense for each taxele to both get and set the field it represents,
which will be necessary in order to allow session behaviour for sources.